remove unused use statements and a bit of cs
