+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2012 Dragonfly Development Inc.
+Copyright (c) 2012-2014 Dragonfly Development Inc.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
