+++ src/Bundle/Command/DumpAutoloadCommand.php
