+++ src/Bundle/Command/InstallCommand.php
