+++ src/Console/Command/UpdateCommand.php
@@ -11,24 +11,24 @@
 
 namespace Dflydev\EmbeddedComposer\Console\Command;
 
-use Composer\Factory;
 use Composer\Installer;
 use Composer\IO\ConsoleIO;
 use Dflydev\EmbeddedComposer\Core\EmbeddedComposerAwareInterface;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * Install Command.
+ * Update Command.
  *
  * @author Jordi Boggiano <j.boggiano@seld.be>
  * @author Ryan Weaver <ryan@knplabs.com>
  * @author Konstantin Kudryashov <ever.zet@gmail.com>
  * @author Beau Simensen <beau@dflydev.com>
  */
-class InstallCommand extends Command
+class UpdateCommand extends Command
 {
     public function __construct($commandPrefix = 'composer:')
     {
@@ -38,11 +38,12 @@ class InstallCommand extends Command
 
     protected function configure()
     {
-        $fullCommand = $this->commandPrefix.'install';
+        $fullCommand = $this->commandPrefix.'update';
         $this
             ->setName($fullCommand)
-            ->setDescription('Install dependencies')
+            ->setDescription('Update dependencies')
             ->setDefinition(array(
+                new InputArgument('packages', InputArgument::IS_ARRAY | InputArgument::OPTIONAL, 'Packages that should be updated, if not provided all packages are.'),
                 new InputOption('prefer-source', null, InputOption::VALUE_NONE, 'Forces installation from package sources when possible, including VCS information.'),
                 new InputOption('dry-run', null, InputOption::VALUE_NONE, 'Outputs the operations but will not execute anything (implicitly enables --verbose).'),
                 new InputOption('dev', null, InputOption::VALUE_NONE, 'Enables installation of dev-require packages.'),
@@ -52,13 +53,20 @@ class InstallCommand extends Command
 The <info>{$fullCommand}</info> command reads a composer.json formatted file.
 The file is read from the current directory unless a project
 directory is specified.
+
+To limit the update operation to a few packages, you can list the package(s)
+you want to update on the command line:
+
+<info>app update vendor/package1 foo/mypackage [...]</info>
+
 EOT
-            );
+            )
+        ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        if (!$this->getApplication() instanceof EmbeddedComposerAwareInterface) {
+        if (!($this->getApplication() instanceof EmbeddedComposerAwareInterface)) {
             throw new \RuntimeException('Application must be instance of EmbeddedComposerAwareInterface');
         }
 
@@ -72,7 +80,10 @@ EOT
             ->setVerbose($input->getOption('verbose'))
             ->setPreferSource($input->getOption('prefer-source'))
             ->setDevMode($input->getOption('dev'))
-            ->setRunScripts(!$input->getOption('no-scripts'));
+            ->setRunScripts(!$input->getOption('no-scripts'))
+            ->setUpdate(true)
+            ->setUpdateWhitelist($input->getArgument('packages'))
+        ;
 
         return $installer->run();
     }
