+++ src/Core/EmbeddedComposerAwareInterface.php
