+++ src/Core/EmbeddedComposerBuilder.php
@@ -15,11 +15,8 @@ use Composer\Autoload\ClassLoader;
 use Composer\Config;
 use Composer\Factory;
 use Composer\Json\JsonFile;
-use Composer\Package\AliasPackage;
-use Composer\Repository\ArrayRepository;
 use Composer\Repository\CompositeRepository;
 use Composer\Repository\InstalledFilesystemRepository;
-use Composer\Repository\FilesystemRepository;
 use Seld\JsonLint\ParsingException;
 
 /**
