+++ src/Core/EmbeddedComposerInterface.php
