+++ src/Core/composer.json
